pyproject.toml
[project]
name = "cozmo-voice-agent"
version = "0.1.0"
description = "Production-ready voice AI agent system handling 100 concurrent PSTN calls"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=0.12,<2.0",
    "livekit-plugins-deepgram>=0.7,<2.0",
    "livekit-plugins-openai>=0.10,<2.0",
    "livekit-plugins-cartesia>=0.4,<2.0",
    "livekit-plugins-silero>=0.7,<2.0",
    "livekit-api>=0.7,<2.0",
    "chromadb>=0.5,<1.0",
    "redis>=5.0,<6.0",
    "prometheus-client>=0.21,<1.0",
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.32,<1.0",
    "websockets>=13,<15",
    "python-dotenv>=1.0,<2.0",
    "psutil>=6.0,<7.0",
    "openai>=1.50,<2.0",
    "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
